modularize employees
